validation added and delete alert added

+++ pages/all-admissions.js
@@ -4,8 +4,25 @@ import { MdDeleteOutline } from "react-icons/md"
 // import Image from 'next/image'
 import { collection, deleteDoc, doc, getDocs } from "firebase/firestore"
 import { fsDb } from "../config/firebase";
-
+import Box from '@mui/material/Box';
+import Typography from '@mui/material/Typography';
+import Modal from '@mui/material/Modal';
+const style = {
+    position: 'absolute',
+    top: '50%',
+    left: '50%',
+    transform: 'translate(-50%, -50%)',
+    width: 400,
+    bgcolor: 'background.paper',
+    boxShadow: 24,
+    p: 4,
+};
 export default function AllAdmissions() {
+    const [open, setOpen] = useState(false);
+    const [delId, setDelId] = useState()
+    const handleOpen = () => setOpen(true);
+    const handleClose = () => setOpen(false);
+
     const memberCollectionRef = collection(fsDb, "members")
     const [check, setCheck] = useState([])
 
@@ -19,11 +36,17 @@ export default function AllAdmissions() {
         }
         getting()
     }, [])
+
     // Delete member 
     const handleDelete = async (item) => {
-        const userDoc = doc(fsDb, "members", item)
-        await deleteDoc(userDoc)
+        if (delId != undefined || delId != "") {
+            const userDoc = doc(fsDb, "members", delId)
+            await deleteDoc(userDoc)
+        }
+        location.reload();
     }
+
+
     return (
         <div className='w-[95%] mt-4 mx-auto'>
             <div className='border-b-[2px] border-[#ffcb04]'>
@@ -55,12 +78,29 @@ export default function AllAdmissions() {
                                 <div className='py-2 border-[1px] border-[transparent] truncate w-[10%] sm:w-[20%] text-center'>{item.member_num}</div>
                                 <div className='py-2 border-[1px] border-[transparent] truncate w-[10%] text-center sm:hidden'>{item.joining_date}</div>
                                 <div className='py-2 border-[1px] border-[transparent] truncate w-[10%] sm:w-[15%] text-center'><Edit item={item} /></div>
-                                <div className='py-2 border-[1px] border-[transparent] truncate w-[10%] sm:w-[15%] text-center'><button onClick={() => handleDelete(item.id)} className='bg-[red] hover:bg-[#ff00001a] hover:text-[red] border-[2px] duration-200 border-[red] px-4 sm:px-2 rounded text-white'><span className='hidden sm:block text-white'><MdDeleteOutline /></span> <span className='sm:hidden'>Delete</span></button></div>
+                                <div className='py-2 border-[1px] border-[transparent] truncate w-[10%] sm:w-[15%] text-center'><button onClick={() => (handleOpen(), setDelId(item.id))} className='bg-[red] hover:bg-[#ff00001a] hover:text-[red] border-[2px] duration-200 border-[red] px-4 sm:px-2 rounded text-white'><span className='hidden sm:block text-white'><MdDeleteOutline /></span> <span className='sm:hidden'>Delete</span></button></div>
                             </div>
                         )
                     })
                 }
             </div>
+            <Modal
+                open={open}
+                onClose={handleClose}
+                aria-labelledby="modal-modal-title"
+                aria-describedby="modal-modal-description"
+            >
+                <Box sx={style}>
+                    <Typography id="modal-modal-title" variant="h5" component="h2" className='text-center underline'>
+                        Delete Member
+                    </Typography>
+                    <Typography id="modal-modal-description" sx={{ m: 2 }} className='text-center'>
+                        Are you sure you want to delete this member ?
+                    </Typography>
+                    <button onClick={() => handleDelete()} className='bg-[red] text-white p-2 mx-2 text-center mx-auto w-[45%] rounded'>Yes</button>
+                    <button onClick={() => handleClose()} className='bg-[#ffcb04] text-white p-2 mx-2 text-center mx-auto w-[45%] rounded'>No</button>
+                </Box>
+            </Modal>
         </div>
     )
 }
